:sparkles: Add delete blockchain application button and use actions hook
